+++ src/thread/executor/future/CallableMainV2.java
@@ -31,7 +31,7 @@ public class CallableMainV2 {
             int value = new Random().nextInt(10);
             log("create value = " + value);
             log("Callable 완료");
-            return value;
+            return (Integer) value;
         }
     }
 }
